# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenHall API",
    "description": "Self-hosted school presence and movement platform API.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-openhall_session",
        "description": "Opaque server-side session bearer (development: openhall_session_dev)."
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Per-session CSRF token from GET /api/v1/auth/session."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/auth/session": {
      "get": {
        "operationId": "getAuthSession",
        "tags": [
          "auth"
        ],
        "description": "Returns the current session. Anonymous callers receive authenticated:false.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "type": "object",
                      "required": [
                        "authenticated"
                      ],
                      "properties": {
                        "authenticated": {
                          "type": "boolean",
                          "enum": [
                            false
                          ]
                        }
                      },
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "required": [
                        "authenticated",
                        "csrfToken",
                        "absoluteExpiresAt",
                        "authenticationMethod"
                      ],
                      "properties": {
                        "authenticated": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        },
                        "csrfToken": {
                          "type": "string",
                          "minLength": 1
                        },
                        "absoluteExpiresAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "authenticationMethod": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "oidc"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "recovery"
                              ]
                            }
                          ]
                        }
                      },
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "operationId": "getMe",
        "tags": [
          "auth"
        ],
        "description": "Minimal canonical identity for the authenticated person (self-only).",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "person",
                    "tenant"
                  ],
                  "properties": {
                    "person": {
                      "type": "object",
                      "required": [
                        "id",
                        "givenName",
                        "familyName",
                        "displayName"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "givenName": {
                          "type": "string"
                        },
                        "familyName": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        }
                      },
                      "additionalProperties": false
                    },
                    "tenant": {
                      "type": "object",
                      "required": [
                        "id",
                        "name",
                        "slug"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "name": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string",
                          "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                          "maxLength": 63
                        }
                      },
                      "additionalProperties": false
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code",
                    "requestId"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[a-z][a-z0-9_]*$"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/discovery": {
      "get": {
        "operationId": "getAuthDiscovery",
        "tags": [
          "auth"
        ],
        "description": "Public login metadata. Returns tenant/provider display info for a single-tenant installation or a known slug; otherwise requires tenant selection without enumerating slugs.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 63
            },
            "in": "query",
            "name": "tenant",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "type": "object",
                      "required": [
                        "tenantSelectionRequired"
                      ],
                      "properties": {
                        "tenantSelectionRequired": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        }
                      },
                      "additionalProperties": false
                    },
                    {
                      "type": "object",
                      "required": [
                        "tenantSelectionRequired",
                        "tenant",
                        "providers"
                      ],
                      "properties": {
                        "tenantSelectionRequired": {
                          "type": "boolean",
                          "enum": [
                            false
                          ]
                        },
                        "tenant": {
                          "type": "object",
                          "required": [
                            "id",
                            "name",
                            "slug"
                          ],
                          "properties": {
                            "id": {
                              "type": "string",
                              "format": "uuid"
                            },
                            "name": {
                              "type": "string"
                            },
                            "slug": {
                              "type": "string",
                              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                              "maxLength": 63
                            }
                          },
                          "additionalProperties": false
                        },
                        "providers": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": [
                              "key",
                              "displayName"
                            ],
                            "properties": {
                              "key": {
                                "type": "string",
                                "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                                "maxLength": 63
                              },
                              "displayName": {
                                "type": "string"
                              }
                            },
                            "additionalProperties": false
                          }
                        }
                      },
                      "additionalProperties": false
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/oidc/callback": {
      "get": {
        "operationId": "completeOidcLogin",
        "tags": [
          "auth"
        ],
        "description": "OIDC callback. Always responds 302: success redirects to the safe local return path with a fresh HttpOnly session cookie; failure redirects to /?error=<code> with a safe machine-readable code. OIDC protections (state/browser binding/nonce/PKCE/replay) apply instead of the SPA CSRF header.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "maxLength": 512
            },
            "in": "query",
            "name": "state",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 4096
            },
            "in": "query",
            "name": "code",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 200
            },
            "in": "query",
            "name": "error",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/auth/recovery": {
      "post": {
        "operationId": "consumeRecoveryGrant",
        "tags": [
          "auth"
        ],
        "description": "Consumes a one-time recovery grant from the Authorization header (never query) and creates a short-lived recovery session.",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "authenticated",
                    "authenticationMethod"
                  ],
                  "properties": {
                    "authenticated": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "authenticationMethod": {
                      "type": "string",
                      "enum": [
                        "recovery"
                      ]
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired recovery token",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code",
                    "requestId"
                  ],
                  "properties": {
                    "type": {
                      "type": "string",
                      "format": "uri-reference"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer",
                      "minimum": 400,
                      "maximum": 599
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[a-z][a-z0-9_]*$"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ]
}